+++ core/views.py
@@ -103,6 +103,7 @@ class Wizard(FormSessionMixin, NamedUrlSessionWizardView):
             data.update(form.cleaned_data)
         del data['terms_agreed']
         del data['captcha']
+        del data['product_name']
         return data
 
 
@@ -117,6 +118,23 @@ class CompaniesHouseSearchAPIView(GenericAPIView):
         response = ch_search_api_client.company.search_companies(query=serializer.validated_data['term'])
         response.raise_for_status()
         return Response(response.json()['items'])
+
+
+class CommodityCodeSearchAPIView(GenericAPIView):
+    serializer_class = serializers.CommodityCodeSearchSerializer
+    permission_classes = []
+    authentication_classes = []
+
+    def get(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.GET)
+        serializer.is_valid(raise_exception=True)
+        response = helpers.lookup_commodity_code_by_name(query=serializer.validated_data['term'])
+        response.raise_for_status()
+        results = [
+            {'text': result['description'], 'value': result['commodity_code']}
+            for result in response.json()['results']
+        ]
+        return Response(results)
 
 
 class SaveForLaterFormView(SuccessMessageMixin, FormView):