+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Pre-release
 
+- TT-1865 - Add product search
 
 ### Implemented enhancements